requirements.txt
cycler==0.11.0
greenlet==1.1.2
importlib-metadata==4.8.3
joblib==1.1.0
kiwisolver==1.3.1
matplotlib==3.3.4
numpy==1.18.5
pandas==1.1.5
Pillow==8.4.0
pyparsing==3.0.7
python-dateutil==2.8.2
python-dotenv~=0.21.0
pytz==2022.1
scikit-learn==0.24.2
scipy==1.5.4
seaborn==0.11.2
six==1.16.0
SQLAlchemy~=1.4.44
threadpoolctl==3.1.0
torch==1.10.2
typing_extensions==4.1.1
zipp==3.6.0
redis==4.4.0
requests==2.28.1
tqdm==4.64.1
beautifulsoup4~=4.11.1
click~=8.1.3
setuptools~=65.6.3
